+++ src/Android/Xamarin.Android/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.jpg
