+++ src/Forms/Shared/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.xaml
@@ -1,7 +1,7 @@
 ﻿<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
-             x:Class="ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster.CreateTerrainSurfaceFromRaster">
+             x:Class="ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceRaster.CreateTerrainSurfaceRaster">
     <Grid>
         <esriUI:SceneView x:Name="MySceneView" />
     </Grid>
