+++ src/Forms/Shared/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.xaml.cs
@@ -1,4 +1,4 @@
-// Copyright 2019 Esri.
+﻿// Copyright 2019 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
@@ -7,16 +7,12 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
 // language governing permissions and limitations under the License.
 
-using Android.App;
-using Android.OS;
-using Android.Widget;
 using ArcGISRuntime.Samples.Managers;
 using Esri.ArcGISRuntime.Mapping;
-using Esri.ArcGISRuntime.UI.Controls;
+using Xamarin.Forms;
 
-namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster
+namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceRaster
 {
-    [Activity (ConfigurationChanges=Android.Content.PM.ConfigChanges.Orientation | Android.Content.PM.ConfigChanges.ScreenSize)]
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         name: "Create terrain surface from a local raster",
         category: "Map",
@@ -24,25 +20,18 @@ namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster
         instructions: "When loaded, the sample will show a scene with a terrain surface applied. Pan and zoom to explore the scene and observe how the terrain surface allows visualizing elevation differences.",
         tags: new[] { "3D", "elevation", "raster", "surface", "terrain" })]
     [ArcGISRuntime.Samples.Shared.Attributes.OfflineData("98092369c4ae4d549bbbd45dba993ebc")]
-    public class CreateTerrainSurfaceFromRaster : Activity
+    public partial class CreateTerrainSurfaceRaster : ContentPage
     {
-        // Hold references to the UI controls.
-        private SceneView _mySceneView;
-
-        protected override void OnCreate(Bundle bundle)
+        public CreateTerrainSurfaceRaster()
         {
-            base.OnCreate(bundle);
-
-            Title = "Create terrain surface from a local raster";
-
-            CreateLayout();
+            InitializeComponent();
             Initialize();
         }
 
         private void Initialize()
         {
             // Create the scene.
-            _mySceneView.Scene = new Scene(Basemap.CreateImagery());
+            MySceneView.Scene = new Scene(Basemap.CreateImagery());
 
             // Get the path to the elevation raster.
             string packagePath = DataManager.GetDataFolder("98092369c4ae4d549bbbd45dba993ebc", "MontereyElevation.dt2");
@@ -57,24 +46,11 @@ namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster
             elevationSurface.ElevationSources.Add(elevationSource);
 
             // Add the surface to the scene.
-            _mySceneView.Scene.BaseSurface = elevationSurface;
+            MySceneView.Scene.BaseSurface = elevationSurface;
 
             // Set an initial camera viewpoint.
             Camera camera = new Camera(36.51, -121.80, 300.0, 0, 70.0, 0.0);
-            _mySceneView.SetViewpointCamera(camera);
-        }
-
-        private void CreateLayout()
-        {
-            // Create a new vertical layout for the app.
-            var layout = new LinearLayout(this) {Orientation = Orientation.Vertical};
-
-            // Add the map view to the layout.
-            _mySceneView = new SceneView(this);
-            layout.AddView(_mySceneView);
-
-            // Show the layout in the app.
-            SetContentView(layout);
+            MySceneView.SetViewpointCamera(camera);
         }
     }
 }
