+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.jpg
