+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.UWP.Samples.CreateTerrainSurfaceFromRaster.CreateTerrainSurfaceFromRaster"
+    x:Class="ArcGISRuntime.UWP.Samples.CreateTerrainSurfaceRaster.CreateTerrainSurfaceRaster"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
