+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.CreateTerrainSurfaceFromRaster.CreateTerrainSurfaceFromRaster"
+﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.CreateTerrainSurfaceRaster.CreateTerrainSurfaceRaster"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
