+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.xaml.cs
@@ -3,15 +3,14 @@
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an 
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 // language governing permissions and limitations under the License.
 
 using ArcGISRuntime.Samples.Managers;
 using Esri.ArcGISRuntime.Mapping;
-using Xamarin.Forms;
 
-namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster
+namespace ArcGISRuntime.WPF.Samples.CreateTerrainSurfaceRaster
 {
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         name: "Create terrain surface from a local raster",
@@ -20,9 +19,9 @@ namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromRaster
         instructions: "When loaded, the sample will show a scene with a terrain surface applied. Pan and zoom to explore the scene and observe how the terrain surface allows visualizing elevation differences.",
         tags: new[] { "3D", "elevation", "raster", "surface", "terrain" })]
     [ArcGISRuntime.Samples.Shared.Attributes.OfflineData("98092369c4ae4d549bbbd45dba993ebc")]
-    public partial class CreateTerrainSurfaceFromRaster : ContentPage
+    public partial class CreateTerrainSurfaceRaster
     {
-        public CreateTerrainSurfaceFromRaster()
+        public CreateTerrainSurfaceRaster()
         {
             InitializeComponent();
             Initialize();
