+++ src/iOS/Xamarin.iOS/Samples/Scene/CreateTerrainSurfaceRaster/CreateTerrainSurfaceRaster.jpg
